+++ extensions/git/src/utils/language-detector.ts
@@ -7,6 +7,8 @@ const MAIN_LANGUAGE_PERCENTAGE = 70;
 const LANGUAGES_COLORS: Record<string, Image.ImageLike> = {
   JavaScript: "https://api.iconify.design/vscode-icons/file-type-js-official.svg",
   TypeScript: "https://api.iconify.design/vscode-icons/file-type-typescript-official.svg",
+  HTML: "https://api.iconify.design/vscode-icons/file-type-html.svg",
+  Markdown: "https://api.iconify.design/simple-icons/markdown.svg",
   Python: "https://api.iconify.design/vscode-icons/file-type-python.svg",
   Java: "https://api.iconify.design/skill-icons/java-light.svg",
   "C#": { source: "https://api.iconify.design/teenyicons/c-sharp-solid.svg", tintColor: Color.Purple },
@@ -23,6 +25,8 @@ const LANGUAGES_COLORS: Record<string, Image.ImageLike> = {
 };
 
 const LANGUAGE_EXTENSION: Record<string, string> = {
+  html: "HTML",
+  md: "Markdown",
   ts: "TypeScript",
   tsx: "TypeScript",
   js: "JavaScript",